feat: Add color luminosity in traceRaySphere() function

+++ include/colors.h
@@ -1,7 +1,7 @@
 #ifndef COLORS_H
 #define COLORS_H
 
-void setLuminosity(uint32_t* myColor, int intensity);
+void setLuminosity(uint32_t* myColor, float intensity);
 uint32_t addColors(uint32_t myColor1, uint32_t myColor2);
 
 #endif
